Add ExploreMapCard, PopularCategoriesSection, and RecentSearchesSection components; integrate them into SearchScreen for enhanced user experience
